refactoring: search performance plateau is now a libary function.
